fix(android): derive background accessibility hiding from sheet stack state
Per-sheet importance snapshots broke with stacked sheets: a parent could
hide the sheet above it, and out-of-order restores could leave app content
permanently hidden. Move hide/restore into TrueSheetStackManager as an
idempotent recompute — only the topmost dimmed sheet (and sheets above it)
stay accessible.
